Fix testing module dependency on httpx
